requirements.txt
asgiref==3.11.1
Django==6.0.2
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
dotenv==0.9.9
package_name==0.1
python-dotenv==1.2.2
sqlparse==0.5.5
